test(rating): recategorize and clean up browser tests
